refactor(workloads/single-module/src/main/kotlin/*): changed package
