+++ workloads/single-module/src/main/kotlin/corp/tbm/cleanwizard/workloads/singlemodule/data/remote/dtos/evolution/EvolutionChainDTOSchema.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanwizard.workloads.multimodule.data.remote.dtos.evolution
+package corp.tbm.cleanwizard.workloads.singlemodule.data.remote.dtos.evolution
 
 import corp.tbm.cleanwizard.foundation.annotations.DTO
 import kotlinx.collections.immutable.ImmutableList
