+++ workloads/single-module/src/main/kotlin/corp/tbm/cleanwizard/workloads/singlemodule/data/remote/dtos/pokemon/PokemonDTOSchema.kt
@@ -1,7 +1,7 @@
-package corp.tbm.cleanwizard.workloads.multimodule.data.remote.dtos.pokemon
+package corp.tbm.cleanwizard.workloads.singlemodule.data.remote.dtos.pokemon
 
 import corp.tbm.cleanwizard.foundation.annotations.DTO
-import corp.tbm.cleanwizard.workloads.multimodule.data.remote.dtos.foundation.PokemonReferenceDTOSchema
+import corp.tbm.cleanwizard.workloads.singlemodule.data.foundation.PokemonReferenceDTOSchema
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.serialization.SerialName
