+++ workloads/single-module/src/main/kotlin/corp/tbm/cleanwizard/workloads/singlemodule/data/remote/dtos/pokemon/PokemonItemDTOSchema.kt
@@ -1,11 +1,10 @@
-package corp.tbm.cleanwizard.workloads.multimodule.data.remote.dtos.pokemon
+package corp.tbm.cleanwizard.workloads.singlemodule.data.remote.dtos.pokemon
 
 import corp.tbm.cleanwizard.foundation.annotations.DTO
-import corp.tbm.cleanwizard.workloads.multimodule.data.remote.dtos.foundation.PokemonReferenceDTOSchema
+import corp.tbm.cleanwizard.workloads.singlemodule.data.foundation.PokemonReferenceDTOSchema
 
 @DTO
 data class PokemonItemDTOSchema(
-    val item : PokemonReferenceDTOSchema,
-    val itemVersionDetails : PokemonItemVersionDetailsDTOSchema,
-
-    )
+    val item: PokemonReferenceDTOSchema,
+    val itemVersionDetails: PokemonItemVersionDetailsDTOSchema,
+)
