+++ workloads/single-module/src/main/kotlin/corp/tbm/cleanwizard/workloads/singlemodule/domain/FooRepository.kt
@@ -1,0 +1,8 @@
+package corp.tbm.cleanwizard.workloads.singlemodule.domain
+
+import corp.tbm.cleanwizard.foundation.annotations.Repository
+
+@Repository
+interface FooRepository {
+    fun debik(lox: List<String>): List<Int>
+}
